requirements.txt

django-import-export
djangorestframework
django
jwt
bcrypt
cryptography
django-log-request-id
